+++ examples/cerrando-canales/cerrando-canales.sh
@@ -1,0 +1,11 @@
+# La idea de cerrar canales nos lleva naturalmente a
+# nuestro siguiente ejemplo: Usar `range` con canales.
+$ go run cerrando-canales.go 
+trabajo enviado 1
+trabajo recibido 1
+trabajo enviado 2
+trabajo recibido 2
+trabajo enviado 3
+trabajo recibido 3
+todos los trabajos han sido enviados
+todos los trabajos han sido recibidos
