+++ examples/cerrando-canales/closing-channels.hash
