+++ examples/declaraciones-cortas/declaraciones-cortas.go
@@ -1,0 +1,9 @@
+package main
+
+import "fmt"
+
+func main() {
+    // `x := val` es una abreviación de `var x type = val`.
+    x := "Hola var"
+    fmt.Println(x)
+}
