+++ examples/declaraciones-cortas/declaraciones-cortas.sh
@@ -1,0 +1,2 @@
+$ go run declaraciones-cortas.go
+Hola var
